fixed logo overlapping buttons when zooming in
When we zoom in the page, the logo overlaps the other buttons on the side menu. Now when you zoom in, the logo will stay in the bottom and if you can't see it anymore, a scroll bar will be appear to go down to the logo. No overlapping occurs anymore.